refactored folder name for user inputs
