+++ src/components/user-input/SearchBar.vue
